refactor: Rename Op to ConnectOp
