+++ src/main/java/com/example/pitching/voice/operation/code/ConnectResOp.java
@@ -1,9 +1,10 @@
-package com.example.pitching.voice.event.op;
+package com.example.pitching.voice.operation.code;
 
+import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.Getter;
 
 @Getter
-public enum ResOp {
+public enum ConnectResOp {
     DISPATCH(0, "Response", "An event was dispatched."),
     HEARTBEAT(1, "Request/Response", "Fired periodically by the client to keep the connection alive."),
     RECONNECT(7, "Response", "You should attempt to reconnect and resume immediately."),
@@ -11,11 +12,12 @@ public enum ResOp {
     HELLO(10, "Response", "Sent immediately after connecting, contains the heartbeat_interval to use."),
     HEARTBEAT_ACK(11, "Response", "Sent in response to receiving a heartbeat to acknowledge that it has been received.");
 
+    @JsonValue
     private final int code;
     private final String direction;
     private final String description;
 
-    ResOp(int code, String direction, String description) {
+    ConnectResOp(int code, String direction, String description) {
         this.code = code;
         this.direction = direction;
         this.description = description;
